Create delete_table
I don't have it in the form of a function yet. 

It fails on line 39. Index exceeds elements array.
The last portion is supposed to find the columns equal to 1 and save it in a new variable which_are_white. Then it draws a line on the row from the first to last column of 1's. Once we draw this line we can fill in the droplet.