Fix bugs surrounding uigetfile
This commit fixes the bug that caused uigetfile to run when the background was pressed. This commit also fixes the error message that appears when a user presses cancel after uigetfile is called. This commit fixes the bug that causes the applet to go to the background when uigetfile is called.